feat(render-model): swap picker to RenderModel + Adapter path
Move picker build/encode logic from gui.ex to the RenderModel pattern:
model (pre-encoded binary + fingerprint), builder (orchestrates
ProtocolGUI encoding including preview), adapter encoder
(fingerprint-based skip). Uses :closed sentinel for dismissed state.
