+++ lib/minga/frontend/adapter/gui.ex
@@ -7,6 +7,7 @@ defmodule Minga.Frontend.Adapter.GUI do
   alias Minga.Frontend.Adapter.GUI.Caches
   alias Minga.Frontend.Adapter.GUI.FileTreeEncoder
   alias Minga.Frontend.Adapter.GUI.GitStatusEncoder
+  alias Minga.Frontend.Adapter.GUI.PickerEncoder
   alias Minga.Frontend.Adapter.GUI.NotificationsEncoder
   alias Minga.Frontend.Adapter.GUI.ObservatoryEncoder
   alias Minga.Frontend.Adapter.GUI.SearchStateEncoder
@@ -84,6 +85,11 @@ defmodule Minga.Frontend.Adapter.GUI do
         do: FileTreeEncoder.encode(ui.file_tree, caches),
         else: {nil, caches}
 
+    {picker_cmd, caches} =
+      if ui.picker,
+        do: PickerEncoder.encode(ui.picker, caches),
+        else: {nil, caches}
+
     cmds =
       Enum.reject(
         [
@@ -100,7 +106,8 @@ defmodule Minga.Frontend.Adapter.GUI do
           tab_bar_cmd,
           workspaces_cmd,
           sidebars_cmd,
-          file_tree_cmd
+          file_tree_cmd,
+          picker_cmd
         ],
         &is_nil/1
       )
